:recycle: refactor: error 폴더 components 밑으로 옮김
