+++ frontend/src/components/error/AuthError.ts
