+++ frontend/src/components/error/AuthErrorBoundary.tsx
