+++ frontend/src/pages/index.tsx
@@ -13,10 +13,10 @@ import Header from '@components/common/Header';
 import RootTitle from '@components/common/Header/RootTitle';
 import NavigationTab from '@components/common/NavigationTab';
 import PageLayout from '@components/common/PageLayout';
+import ApiErrorBoundary from '@components/error/ApiErrorBoundary';
 import { Category, CategoryKr } from '@constants/category';
 import { Location, LocationKr } from '@constants/location';
 import { PAGE_TITLE } from '@constants/pageTitle';
-import ApiErrorBoundary from '@error/ApiErrorBoundary';
 
 const Main = () => {
   const {
